refactor: User 정밀도 컬럼 BigDecimal로 변환 (height/weight/target)

+++ backend/src/main/java/com/ssafy/s309/domain/user/dto/SettingsResponse.java
@@ -2,6 +2,7 @@ package com.ssafy.s309.domain.user.dto;
 
 import com.ssafy.s309.domain.user.entity.DiabetesType;
 import com.ssafy.s309.domain.user.entity.User;
+import java.math.BigDecimal;
 
 public record SettingsResponse(
     Integer userId,
@@ -9,12 +10,12 @@ public record SettingsResponse(
     Integer age,
     String gender,
     String phone,
-    Float height,
-    Float weight,
+    BigDecimal height,
+    BigDecimal weight,
     DiabetesType diabetesType,
     Boolean isMedicated,
-    Integer targetLow,
-    Integer targetHigh,
+    BigDecimal targetLow,
+    BigDecimal targetHigh,
     Integer weekStartDay) {
 
   public static SettingsResponse from(User user) {
