+++ backend/src/main/java/com/ssafy/s309/domain/user/dto/SettingsUpdateRequest.java
@@ -1,16 +1,17 @@
 package com.ssafy.s309.domain.user.dto;
 
 import com.ssafy.s309.domain.user.entity.DiabetesType;
+import java.math.BigDecimal;
 
 public record SettingsUpdateRequest(
     String name,
     Integer age,
     String gender,
     String phone,
-    Float height,
-    Float weight,
+    BigDecimal height,
+    BigDecimal weight,
     DiabetesType diabetesType,
     Boolean isMedicated,
-    Integer targetLow,
-    Integer targetHigh,
+    BigDecimal targetLow,
+    BigDecimal targetHigh,
     Integer weekStartDay) {}
